+++ src/gallery/image_processing/tests/composite_tests.rs
@@ -1,0 +1,1 @@
+// TODO: Move composite image tests from image_processing_old.rs