+++ src/gallery/image_processing/tests/jpeg_tests.rs
@@ -1,0 +1,1 @@
+// TODO: Move JPEG-specific tests from image_processing_old.rs