+++ src/gallery/image_processing/tests/png_tests.rs
@@ -1,0 +1,1 @@
+// TODO: Move PNG-specific tests from image_processing_old.rs